% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefit.R
\name{dung_scenarios}
\alias{dung_scenarios}
\title{The three dung beetle abundance scenarios}
\usage{
dung_scenarios()
}
\value{
Character vector of the valid \code{scenario} labels.
}
\description{
The three dung beetle abundance scenarios
}
